:recycle: StreamController 를 사용하여 변경
